Basic C#

+++ NetCentricComputing/NetCentricComputing/FileReadWrite.cs
@@ -10,8 +10,8 @@ namespace NetCentricComputing
         static void Main(string[] args)
         {
             string path = "file1.txt";
-            string newPath = "file2.txt";
-            if (File.Exists(path))
+            string newPath = "file3.txt";
+            if (!File.Exists(path))
             {
                 File.WriteAllText(path, "This is a file handling in c#.");
             }
@@ -20,8 +20,15 @@ namespace NetCentricComputing
             }
             File.AppendAllText(path,"Appending in the new line in the existing file");
             string mytxt = File.ReadAllText(path);
-            
-            File.Copy(path, newPath);
+            if (!File.Exists(newPath))
+            {
+                File.Copy(path, newPath);
+            }
+            else
+            {
+                Console.WriteLine($"File already exists in {newPath}");
+
+            }
             string newtxt = File.ReadAllText(newPath);
             Console.WriteLine(newtxt);
             Console.ReadKey();
